requirements.txt
streamlit>=1.32.0
yfinance==0.2.36
requests==2.31.0
python-dotenv==1.0.1
sendgrid==6.10.0 